+++ encore_api_cli/utils.py
@@ -1,15 +1,15 @@
 import click
 
 from encore_api_cli.client import Client
-from encore_api_cli.config import Config
+from encore_api_cli.settings import Settings
 
 
 def get_client(profile):
-    config = Config(profile)
-    if not config.is_ok:
+    settings = Settings(profile)
+    if not settings.is_ok():
         message = ' '.join([
             'The credentials is invalid or not set.',
             'Run "encore configure" to set credentials.'
         ])
         raise click.ClickException(message)
-    return Client(config.client_id, config.client_secret, config.url)
+    return Client(settings.client_id, settings.client_secret, settings.url)